chore(daemon): remove legacy SolarAutomation state machine and pool.solar.* config (#406)
Final cleanup step of the #401 solar-automation redesign, landed after #402
(on-device runtime accumulator), #403 (daemon solar-aggregator publish), and
#405 (on-device solar hysteresis) are already deployed and verified.

Deletes myhome/daemon/solar_automation.go (SolarConfig, PumpController,
RuntimeTracker, SolarAutomation, shellyPumpController, computeRuntimeTargets)
and its tests, plus myhome/daemon/pool_runtime_tracker.go and its test (no
remaining callers after #402 moved runtime/turnover tracking on-device).
readPoolKVSFloat/readPoolKVSInt, still needed by pool_notices.go's
ComputeTurnover, move there instead of being deleted. Removes the daemon
wiring block that constructed shellyPumpController/SolarAutomation, and the
seven PoolSolar* options/flags/viper bindings (PoolEnabled/PoolDeviceID are
kept — they still gate PoolNotices, the claimers registry, and
pool.getstatus). digest.go drops the now-dead pool.solar_start/pool.solar_stop
cases; solar-triggered starts/stops already flow through pool.pump_start/stop
with a descriptive reason since #405.

Updates docs/configuration.md, myhome-example.yaml, docs/beem-energy.md
(Parts 2-3 rewritten with a revision note explaining the reversed Option
A/B decision), and docs/notice-events-plan.md to match.

Verified: an existing myhome.yaml with a pool.solar: block (e.g. the NAS's
deployed v0.11.9 config) loads without error or warning — viper simply
never reads those now-unbound keys, so they become inert. No migration
code needed.
