+++ docs/configuration.md
@@ -555,7 +555,12 @@ Credentials for the SFR home gateway. Authentication is skipped when either valu
 
 ## Pool
 
-The pool runtime tracker reports how many seconds the pool pump has run today by querying the shared events database (`events.db`). The gen2 listener already captures every switch ON/OFF event from all Shelly devices — no separate pool database is needed.
+`pool-pump.js` itself tracks how many seconds it has run today and how much water-volume turnover
+it has achieved, mirroring both to KVS (`script/pool-pump/runtime-sec`, `script/pool-pump/turnover-today`)
+as it starts/stops (see `docs/beem-energy.md` "Part 3"). The daemon's `PoolNotices` type
+(`myhome/daemon/pool_notices.go`) just reads those KVS values back — no separate pool database, and
+no dependency on the shared events database — to power `ctl pool status`, the web UI's pool tags,
+and the `pool.turnover_today` notice recorded whenever the pump stops.
 
 ### Example
 
@@ -572,54 +577,19 @@ pool:
 | `pool.device_id` | `MYHOME_POOL_DEVICE_ID` | `--pool-device-id` | — | Pool Shelly device ID (e.g. `shellyplus1pm-aabbccddeeff`) |
 | `pool.enabled` | `MYHOME_POOL_ENABLED` | `--enable-pool` | `false` | Enable pool runtime tracking |
 
-### Solar automation
+### Solar-driven pool pump
 
-The solar automation goroutine subscribes to Beem Energy power samples and controls the pool pump using a hysteresis state machine:
+Solar-driven pump control no longer lives in the daemon. As of the #401 redesign, `pool-pump.js`
+itself subscribes to the daemon-published `myhome/energy/solar/available` topic (see
+`docs/beem-energy.md` "Part 2") and decides for itself whether to start/stop, using its own
+KVS-configured hysteresis thresholds (`script/pool-pump/solar-*` keys — see `docs/pool-pump.md`)
+and its own existing `doStart`/`doStop`. This keeps the anti-cycling fuse, `isMyTurnToRun()`, and
+water-supply protection in force for solar-triggered runs exactly as for scheduled/manual runs, and
+lets the pump keep working on its forecast-derived schedule when the daemon is down.
 
-- **IDLE → RUNNING** when `solar_w ≥ start_threshold_w` for `start_delay` (and the hard ceiling hasn't been reached today)
-- **RUNNING → IDLE** when the hard ceiling (`max_volume_turnover`) is reached — always, regardless of solar
-- **RUNNING → IDLE** when the soft-stop target (`min_volume_turnover`) is reached **and** `solar_w < start_threshold_w` — while solar is still producing, the pump keeps running past the soft target to use free energy
-- **RUNNING → IDLE** when `solar_w < stop_threshold_w` for `stop_delay`
-
-`min_volume_turnover` and `max_volume_turnover` are dimensionless multipliers (pool volumes filtered per day). At startup the daemon reads `script/pool-pump/{pool-volume,max-flow-rate,max-rpm,speed}` from the pool device KVS — the same values pool-pump.js uses for its own scheduling — and derives `daily_target_sec` / `max_rotation_sec`:
-
-```
-flow_rate        = max_flow_rate × (speed / max_rpm)
-daily_target_sec = pool_volume × min_volume_turnover / flow_rate × 3600
-max_rotation_sec = pool_volume × max_volume_turnover / flow_rate × 3600
-```
-
-The daemon only reads these KVS keys, never writes them — KVS remains exclusively the JS script's domain. Solar automation is disabled (with a logged error) if `max_volume_turnover < min_volume_turnover` or if any of the four KVS keys is missing or non-numeric.
-
-Requires both `pool.device_id` and Beem Energy credentials (`beem.email` + `beem.password`) to be configured.
-
-#### Example
-
-```yaml
-pool:
-  device_id: "shellyplus1pm-aabbccddeeff"
-  enabled: true
-  solar:
-    enabled: true
-    start_threshold_w:   500
-    stop_threshold_w:    200
-    start_delay:         5m
-    stop_delay:          10m
-    min_volume_turnover: 5   # soft stop: stop once filtered AND solar gone
-    max_volume_turnover: 7   # hard ceiling: always stop once filtered
-```
-
-#### Options
-
-| Key | Env var | Flag | Default | Description |
-|-----|---------|------|---------|-------------|
-| `pool.solar.enabled` | `MYHOME_POOL_SOLAR_ENABLED` | `--enable-pool-solar` | `false` | Enable solar-driven pump automation |
-| `pool.solar.start_threshold_w` | `MYHOME_POOL_SOLAR_START_THRESHOLD_W` | `--pool-solar-start-threshold-w` | `500` | Solar power threshold to start pump (W) |
-| `pool.solar.stop_threshold_w` | `MYHOME_POOL_SOLAR_STOP_THRESHOLD_W` | `--pool-solar-stop-threshold-w` | `200` | Solar power threshold to stop pump (W) |
-| `pool.solar.start_delay` | `MYHOME_POOL_SOLAR_START_DELAY` | `--pool-solar-start-delay` | `5m` | Solar must hold above start threshold for this long |
-| `pool.solar.stop_delay` | `MYHOME_POOL_SOLAR_STOP_DELAY` | `--pool-solar-stop-delay` | `10m` | Solar must hold below stop threshold for this long |
-| `pool.solar.min_volume_turnover` | `MYHOME_POOL_SOLAR_MIN_VOLUME_TURNOVER` | `--pool-solar-min-volume-turnover` | `5` | Soft-stop target: pool volumes filtered per day; pump keeps running past this while solar is still above the start threshold |
-| `pool.solar.max_volume_turnover` | `MYHOME_POOL_SOLAR_MAX_VOLUME_TURNOVER` | `--pool-solar-max-volume-turnover` | `7` | Hard ceiling: pool volumes filtered per day; pump always stops (and won't be solar-started) once reached |
+The daemon's only remaining role here is publishing the aggregate solar wattage (see "Solar
+Aggregator" above) and reading back the on-device runtime/turnover accumulators described above for
+`ctl pool status` and the `pool.turnover_today` notice.
 
 ## Notice & Email (SMTP)
