+++ myhome-example.yaml
@@ -161,7 +161,16 @@ events:
   # enabled: true
 
 # Pool Runtime Tracking Configuration
-# Pool runtime is derived from the shared events database (events.db) — no separate pool.db.
+# pool-pump.js tracks its own daily runtime/turnover and mirrors it to KVS
+# (script/pool-pump/runtime-sec, script/pool-pump/turnover-today); the daemon
+# just reads those values back for `ctl pool status` and the
+# pool.turnover_today notice — no separate pool database.
+#
+# Solar-driven pump control also lives entirely in pool-pump.js now (it
+# subscribes to myhome/energy/solar/available, published below by the solar
+# aggregator, and applies its own KVS-configured hysteresis — see
+# docs/pool-pump.md for the script/pool-pump/solar-* keys). There is no
+# daemon-side pool.solar.* config anymore.
 pool:
   # Shelly device ID for the pool pump
   # Example: shellyplus1pm-aabbccddeeff
@@ -170,43 +179,6 @@ pool:
   # Set to true to enable pool runtime tracking
   # Default: false
   # enabled: false
-
-  # Solar-driven pump automation
-  solar:
-    # Set to true to enable solar-driven pump control
-    # Default: false
-    # enabled: false
-
-    # Start pump when solar production exceeds this value (W)
-    # Default: 500
-    # start_threshold_w: 500
-
-    # Stop pump when solar production falls below this value (W)
-    # Default: 200
-    # stop_threshold_w: 200
-
-    # Solar must hold above start_threshold_w for this long before the pump starts
-    # Default: 5m
-    # start_delay: 5m
-
-    # Solar must hold below stop_threshold_w for this long before the pump stops
-    # Default: 10m
-    # stop_delay: 10m
-
-    # Soft-stop target: pool volumes filtered per day. Reaching this only stops
-    # the pump once solar has also dropped below start_threshold_w — while solar
-    # is still producing, the pump keeps running past this target to use free energy.
-    # Converted to daily_target_sec at startup via pool device KVS
-    # (script/pool-pump/{pool-volume,max-flow-rate,max-rpm,speed}).
-    # Default: 5
-    # min_volume_turnover: 5
-
-    # Hard ceiling: pool volumes filtered per day. The pump always stops once
-    # reached, regardless of solar, and won't be started via solar past this point.
-    # Converted to max_rotation_sec at startup the same way as min_volume_turnover.
-    # Must be >= min_volume_turnover.
-    # Default: 7
-    # max_volume_turnover: 7
 
 # Beem Energy: set email and password to enable solar production polling.
 # Integration is enabled automatically when both values are non-empty.
