+++ myhome/ctl/options/options.go
@@ -95,13 +95,6 @@ var Flags struct {
 	SolarStaleAfter             time.Duration // solar aggregator: a source's last reading older than this is excluded from the sum
 	SFRUsername                 string        // SFR box account username; from .env, never a flag
 	SFRPassword                 string        // SFR box account password; from .env, never a flag
-	PoolSolarEnabled            bool          // whether to enable solar-driven pool pump automation
-	PoolSolarStartThresholdW    float64       // solar power threshold to start pump (W)
-	PoolSolarStopThresholdW     float64       // solar power threshold to stop pump (W)
-	PoolSolarStartDelay         time.Duration // solar must hold above start threshold for this long
-	PoolSolarStopDelay          time.Duration // solar must hold below stop threshold for this long
-	PoolSolarMinVolumeTurnover  float64       // soft-stop target: pool volumes filtered per day (converted to daily_target_sec via pool KVS)
-	PoolSolarMaxVolumeTurnover  float64       // hard ceiling: pool volumes filtered per day (converted to max_rotation_sec via pool KVS)
 	EnableNoticeService         bool          // whether to enable the notice service (motion rule + daily email digest)
 	NoticeNightStart            string        // "HH:MM" start of the night window used by the motion rule
 	NoticeNightEnd              string        // "HH:MM" end of the night window used by the motion rule
